6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int b;
 float c;
 long int d;
 long long int e;
-printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu bytes(s)\n", (unsigned long) sizeof(b));
-printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
-printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(e));
-printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(c));
+printf("Size of a char: %lu bytes(s)\n", sizeof(a));
+printf("Size of an int: %lu bytes(s)\n", sizeof(b));
+printf("Size of a long int: %lu bytes(s)\n", sizeof(d));
+printf("Size of a long long int: %lu bytes(s)\n", sizeof(e));
+printf("Size of a float: %lu bytes(s)\n", sizeof(c));
 return (0);
 }
